reverted workflow embeddings. Cleaned up code
