+++ docs/reference/dataset/index.md
@@ -8,6 +8,3 @@
 ::: kolena.dataset.dataset
 ::: kolena.dataset.evaluation
 ::: kolena.dataset.search
-    options:
-        show_root_heading: false
-        members: ["upload_dataset_embeddings"]
